[FLINK-38211][table] Move testMultiSinkOnMultiJoinedView to MJ tests
