+++ flink-table/flink-table-planner/src/test/java/org/apache/flink/table/planner/plan/stream/sql/MultiJoinTest.java
@@ -18,18 +18,25 @@
 
 package org.apache.flink.table.planner.plan.stream.sql;
 
+import org.apache.flink.table.api.ExplainDetail;
+import org.apache.flink.table.api.StatementSet;
 import org.apache.flink.table.api.TableConfig;
 import org.apache.flink.table.api.config.OptimizerConfigOptions;
+import org.apache.flink.table.planner.utils.PlanKind;
+import org.apache.flink.table.planner.utils.StreamTableTestUtil;
 import org.apache.flink.table.planner.utils.TableTestBase;
-import org.apache.flink.table.planner.utils.TableTestUtil;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import scala.Enumeration;
+
+import static scala.runtime.BoxedUnit.UNIT;
+
 /** Tests for multi-join plans. */
 public class MultiJoinTest extends TableTestBase {
 
-    private TableTestUtil util;
+    private StreamTableTestUtil util;
 
     @BeforeEach
     void setup() {
@@ -552,5 +559,90 @@ public class MultiJoinTest extends TableTestBase {
                         + "  ON  u.user_id = p.user_id AND p.price >= 0 "
                         + "JOIN AddressPK a"
                         + "  ON  u.user_id = a.user_id AND a.location IS NOT NULL");
+    }
+
+    @Test
+    void testMultiSinkOnMultiJoinedView() {
+        util.tableEnv()
+                .executeSql(
+                        "create temporary table src1 (\n"
+                                + "  a int,\n"
+                                + "  b bigint,\n"
+                                + "  c string,\n"
+                                + "  d int,\n"
+                                + "  primary key(a, c) not enforced\n"
+                                + ") with (\n"
+                                + " 'connector' = 'values',\n"
+                                + " 'changelog-mode' = 'I,UA,UB,D'\n"
+                                + ")");
+        util.tableEnv()
+                .executeSql(
+                        "create temporary table src2 (\n"
+                                + "  a int,\n"
+                                + "  b bigint,\n"
+                                + "  c string,\n"
+                                + "  d int,\n"
+                                + "  primary key(a, c) not enforced\n"
+                                + ") with (\n"
+                                + " 'connector' = 'values',\n"
+                                + " 'changelog-mode' = 'I,UA,UB,D'\n"
+                                + ")");
+        util.tableEnv()
+                .executeSql(
+                        "create temporary table sink1 (\n"
+                                + "  a int,\n"
+                                + "  b string,\n"
+                                + "  c bigint,\n"
+                                + "  d bigint\n"
+                                + ") with (\n"
+                                + " 'connector' = 'values',\n"
+                                + " 'sink-insert-only' = 'false'\n"
+                                + ")");
+        util.tableEnv()
+                .executeSql(
+                        "create temporary table sink2 (\n"
+                                + "  a int,\n"
+                                + "  b string,\n"
+                                + "  c bigint,\n"
+                                + "  d string\n"
+                                + ") with (\n"
+                                + " 'connector' = 'values',\n"
+                                + " 'sink-insert-only' = 'false'\n"
+                                + ")");
+        util.tableEnv()
+                .executeSql(
+                        "create temporary view v1 as\n"
+                                + "select\n"
+                                + "  t1.a as a, t1.`day` as `day`, t2.b as b, t2.c as c\n"
+                                + "from (\n"
+                                + "  select a, b, DATE_FORMAT(CURRENT_TIMESTAMP, 'yyMMdd') as `day`\n"
+                                + "  from src1\n"
+                                + " ) t1\n"
+                                + "join (\n"
+                                + "  select b, CONCAT(c, DATE_FORMAT(CURRENT_TIMESTAMP, 'yyMMdd')) as `day`, c, d\n"
+                                + "  from src2\n"
+                                + ") t2\n"
+                                + " on t1.a = t2.d");
+
+        StatementSet stmtSet = util.tableEnv().createStatementSet();
+        stmtSet.addInsertSql(
+                "insert into sink1\n"
+                        + "  select a, `day`, sum(b), count(distinct c)\n"
+                        + "  from v1\n"
+                        + "  group by a, `day`");
+        stmtSet.addInsertSql(
+                "insert into sink2\n"
+                        + "  select a, `day`, b, c\n"
+                        + "  from v1\n"
+                        + "  where b > 100");
+
+        util.doVerifyPlan(
+                stmtSet,
+                new ExplainDetail[] {ExplainDetail.PLAN_ADVICE},
+                false,
+                new Enumeration.Value[] {PlanKind.OPT_REL_WITH_ADVICE()},
+                () -> UNIT,
+                false,
+                false);
     }
 }
